refactor: map hooks
