+++ src/page-modules/departures/server/journey-planner/index.ts
@@ -43,7 +43,7 @@ import {
   type TransportModeType,
   isTransportModeType,
   filterGraphQlTransportModes,
-  TransportSubmodeType,
+  type TransportSubmodeType,
 } from '@atb/modules/transport-mode';
 import {
   ServiceJourneyWithEstimatedCallsDocument,
